requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
psycopg2-binary==2.9.4
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.9.0
